requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.3.0.23
boto3==1.16.56
botocore==1.19.56
celery==3.1.26.post2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.10
django-admin-export-action==0.2.2
django-admin-interface==0.15.6
django-adminfiles==1.0.1
django-audiofield==0.10.0
django-celery==3.3.1
asgi_redis
django-colorfield==0.4.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.5.0
django-storages==1.11.1
django-summernote==0.8.11.6
djangorestframework==3.12.2
djongo==1.3.3
drf-extra-fields==3.1.1
et-xmlfile==1.0.1
filelock==3.0.12
gTTS==2.2.1
gunicorn==20.0.4
google-cloud-storage
idna==2.10
importlib-metadata==3.7.3
isort==5.6.4
jmespath==0.10.0
kombu==3.0.37
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.7
pep8==1.7.1
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.2
PyMySQL==0.10.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
PyYAML==5.4.1
requests==2.25.0
s3transfer==0.3.4
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.2.4
tablib==3.0.0
toml==0.10.2
Unipath==1.1
urllib3==1.26.2
ve==0.0.1
vine==5.0.0
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
whitelist==0.2.8
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
